refactor: presentation&application 패키지명 변경
