+++ backend/src/main/java/zipgo/common/config/WebConfig.java
@@ -1,4 +1,4 @@
-package zipgo.config;
+package zipgo.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
