+++ backend/src/main/java/zipgo/petfood/presentation/PetFoodController.java
@@ -1,4 +1,4 @@
-package zipgo.petfood.controller;
+package zipgo.petfood.presentation;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import zipgo.petfood.controller.dto.GetPetFoodsResponse;
+import zipgo.petfood.presentation.dto.GetPetFoodsResponse;
 import zipgo.petfood.domain.PetFood;
-import zipgo.petfood.service.PetFoodService;
+import zipgo.petfood.application.PetFoodService;
 
 @RestController
 @RequestMapping("/pet-foods")
