+++ backend/src/test/java/zipgo/petfood/domain/fixture/PetFoodFixture.java
@@ -1,4 +1,4 @@
-package zipgo.domain.fixture;
+package zipgo.petfood.domain.fixture;
 
 import zipgo.petfood.domain.PetFood;
