+++ backend/src/test/java/zipgo/petfood/domain/repository/PetFoodRepositoryTest.java
@@ -1,7 +1,6 @@
-package zipgo.domain.repository;
+package zipgo.petfood.domain.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static zipgo.domain.fixture.PetFoodFixture.키워드_없이_식품_초기화;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
@@ -9,8 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import zipgo.petfood.domain.Keyword;
 import zipgo.petfood.domain.PetFood;
-import zipgo.petfood.domain.repository.KeywordRepository;
-import zipgo.petfood.domain.repository.PetFoodRepository;
+import zipgo.petfood.domain.fixture.PetFoodFixture;
 
 @DataJpaTest
 class PetFoodRepositoryTest {
@@ -24,8 +22,8 @@ class PetFoodRepositoryTest {
     @Test
     void 모든_식품을_조회할_수_있다() {
         // given
-        PetFood 반려동물_식품_1 = petFoodRepository.save(키워드_없이_식품_초기화());
-        PetFood 반려동물_식품_2 = petFoodRepository.save(키워드_없이_식품_초기화());
+        PetFood 반려동물_식품_1 = petFoodRepository.save(PetFoodFixture.키워드_없이_식품_초기화());
+        PetFood 반려동물_식품_2 = petFoodRepository.save(PetFoodFixture.키워드_없이_식품_초기화());
 
         // when
         List<PetFood> petFoods = petFoodRepository.findAll();
@@ -37,7 +35,7 @@ class PetFoodRepositoryTest {
     @Test
     void 키워드로_식품을_조회할_수_있다() {
         // given
-        PetFood 키워드가_없는_식품 = petFoodRepository.save(키워드_없이_식품_초기화());
+        PetFood 키워드가_없는_식품 = petFoodRepository.save(PetFoodFixture.키워드_없이_식품_초기화());
         PetFood 키워드가_있는_식품 = 키워드있는_식품_생성하기();
 
         // when
