+++ backend/src/test/java/zipgo/petfood/service/PetFoodServiceTest.java
@@ -1,11 +1,13 @@
-package zipgo.service;
+package zipgo.petfood.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static zipgo.domain.fixture.PetFoodFixture.키워드_없이_식품_초기화;
+import static zipgo.petfood.domain.fixture.PetFoodFixture.키워드_없이_식품_초기화;
 
 import java.util.List;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -14,19 +16,18 @@ import zipgo.petfood.domain.PetFood;
 import zipgo.petfood.domain.repository.KeywordRepository;
 import zipgo.petfood.domain.repository.PetFoodRepository;
 import zipgo.petfood.exception.KeywordException;
-import zipgo.petfood.service.PetFoodService;
+import zipgo.petfood.application.PetFoodService;
+import zipgo.petfood.stub.FakeKeywordRepository;
+import zipgo.petfood.stub.FakePetFoodRepository;
 
-@SpringBootTest
+@DisplayNameGeneration(ReplaceUnderscores.class)
 class PetFoodServiceTest {
 
-    @Autowired
-    private PetFoodRepository petFoodRepository;
+    private PetFoodRepository petFoodRepository = new FakePetFoodRepository();
 
-    @Autowired
-    private KeywordRepository keywordRepository;
+    private KeywordRepository keywordRepository = new FakeKeywordRepository();
 
-    @Autowired
-    private PetFoodService petFoodService;
+    private PetFoodService petFoodService = new PetFoodService(petFoodRepository, keywordRepository);
 
     @Test
     void 키워드가_다이어트인_식품_목록을_조회한다() {
@@ -45,11 +46,12 @@ class PetFoodServiceTest {
     }
 
     private PetFood 다이어트_키워드_식품_저장() {
-        Keyword 다이어트_키워드 = keywordRepository.findByName("diet").orElseThrow();
+        Keyword keyword = keywordRepository.save(new Keyword("diet"));
         PetFood 다이어트_키워드_식품 = petFoodRepository.save(
                 new PetFood("[고집] 돌아온 배배",
                         "https://github.com/woowacourse-teams/2023-zipgo",
-                        "https://avatars.githubusercontent.com/u/94087228?v=4", 다이어트_키워드));
+                        "https://avatars.githubusercontent.com/u/94087228?v=4",
+                        keyword));
         return 다이어트_키워드_식품;
     }
 
@@ -58,8 +60,7 @@ class PetFoodServiceTest {
         // given
         String 없는_키워드 = "없는 키워드";
 
-        // when
-        // then
+        // when, then
         assertThatThrownBy(() -> petFoodService.getPetFoodHaving(없는_키워드))
                 .isInstanceOf(KeywordException.NotFound.class);
     }
